+++ tests/UnitTests/RawArrayOperations/TestSources/RawArrayOperations.cpp
@@ -35,7 +35,7 @@ LIBRARY_LINK_FUNCTION(echoRawArray) {
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
 
 /*
@@ -55,7 +55,7 @@ LIBRARY_LINK_FUNCTION(getRawArrayLength) {
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
 
 LIBRARY_LINK_FUNCTION(getRawArrayRank) {
@@ -72,7 +72,7 @@ LIBRARY_LINK_FUNCTION(getRawArrayRank) {
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
 
 //create new raw array
@@ -89,7 +89,7 @@ LIBRARY_LINK_FUNCTION(newRawArray) {
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
 
 //clone RawArray
@@ -108,7 +108,7 @@ LIBRARY_LINK_FUNCTION(cloneRawArray) {
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
 
 LIBRARY_LINK_FUNCTION(changeSharedRawArray) {
@@ -162,5 +162,5 @@ EXTERN_C DLLEXPORT int rawZeroData(WolframLibraryData libData, mint Argc, MArgum
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
